requirements.txt
fastapi
fpdf2
google-generativeai
langchain-community
langchain-google-genai
langchain-qdrant
langchain-text-splitters
openai
pypdf
python-dotenv
redis
rq
uvicorn